initial work and refactor statistics functions
