+++ core/src/main/java/io/questdb/griffin/engine/functions/groupby/StdDevSampleDoubleGroupByFunction.java
@@ -29,10 +29,8 @@ import io.questdb.cairo.ColumnType;
 import io.questdb.cairo.map.MapValue;
 import io.questdb.cairo.sql.Function;
 import io.questdb.cairo.sql.Record;
-import io.questdb.griffin.engine.functions.DoubleFunction;
-import io.questdb.griffin.engine.functions.GroupByFunction;
-import io.questdb.griffin.engine.functions.UnaryFunction;
 import io.questdb.std.Numbers;
+// import io.questdb.std.StatAggregator;
 import org.jetbrains.annotations.NotNull;
 
 /**
@@ -40,27 +38,27 @@ import org.jetbrains.annotations.NotNull;
  *
  * @see <a href="https://en.wikipedia.org/wiki/Algorithms_for_calculating_variance#Welford's_online_algorithm">Welford's algorithm</a>
  */
-public class StdDevSampleDoubleGroupByFunction extends DoubleFunction implements GroupByFunction, UnaryFunction {
-    private final Function arg;
-    private int valueIndex;
+public class StdDevSampleDoubleGroupByFunction extends AbstractStatAggregatorGroupByFunction {
+    // private final StatAggregator statAggregator;
 
     public StdDevSampleDoubleGroupByFunction(@NotNull Function arg) {
-        this.arg = arg;
+        super(arg);
     }
 
     @Override
     public void computeFirst(MapValue mapValue, Record record) {
         final double d = arg.getDouble(record);
         if (Numbers.isFinite(d)) {
-            double mean = 0;
-            double sum = 0;
-            long count = 1;
-            double oldMean = mean;
-            mean += (d - mean) / count;
-            sum += (d - mean) * (d - oldMean);
-            mapValue.putDouble(valueIndex, mean);
-            mapValue.putDouble(valueIndex + 1, sum);
-            mapValue.putLong(valueIndex + 2, 1L);
+            // double mean = 0;
+            // double sum = 0;
+            // long count = 1;
+            // double oldMean = mean;
+            // mean += (d - mean) / count;
+            // sum += (d - mean) * (d - oldMean);
+            // mapValue.putDouble(valueIndex, mean);
+            // mapValue.putDouble(valueIndex + 1, sum);
+            // mapValue.putLong(valueIndex + 2, 1L);
+            aggregate(mapValue, d);
         } else {
             mapValue.putDouble(valueIndex, 0);
             mapValue.putDouble(valueIndex + 1, 0);
@@ -72,21 +70,17 @@ public class StdDevSampleDoubleGroupByFunction extends DoubleFunction implements
     public void computeNext(MapValue mapValue, Record record) {
         final double d = arg.getDouble(record);
         if (Numbers.isFinite(d)) {
-            double mean = mapValue.getDouble(valueIndex);
-            double sum = mapValue.getDouble(valueIndex + 1);
-            long count = mapValue.getLong(valueIndex + 2) + 1;
-            double oldMean = mean;
-            mean += (d - mean) / count;
-            sum += (d - mean) * (d - oldMean);
-            mapValue.putDouble(valueIndex, mean);
-            mapValue.putDouble(valueIndex + 1, sum);
-            mapValue.addLong(valueIndex + 2, 1L);
+            // double mean = mapValue.getDouble(valueIndex);
+            // double sum = mapValue.getDouble(valueIndex + 1);
+            // long count = mapValue.getLong(valueIndex + 2) + 1;
+            // double oldMean = mean;
+            // mean += (d - mean) / count;
+            // sum += (d - mean) * (d - oldMean);
+            // mapValue.putDouble(valueIndex, mean);
+            // mapValue.putDouble(valueIndex + 1, sum);
+            // mapValue.addLong(valueIndex + 2, 1L);
+            aggregate(mapValue, d);
         }
-    }
-
-    @Override
-    public Function getArg() {
-        return arg;
     }
 
     @Override
@@ -103,31 +97,5 @@ public class StdDevSampleDoubleGroupByFunction extends DoubleFunction implements
     @Override
     public String getName() {
         return "stddev_samp";
-    }
-
-    @Override
-    public boolean isConstant() {
-        return false;
-    }
-
-    @Override
-    public void pushValueTypes(ArrayColumnTypes columnTypes) {
-        this.valueIndex = columnTypes.getColumnCount();
-        columnTypes.add(ColumnType.DOUBLE);
-        columnTypes.add(ColumnType.DOUBLE);
-        columnTypes.add(ColumnType.LONG);
-    }
-
-    @Override
-    public void setDouble(MapValue mapValue, double value) {
-        mapValue.putDouble(valueIndex, value);
-        mapValue.putLong(valueIndex + 2, 1L);
-    }
-
-    @Override
-    public void setNull(MapValue mapValue) {
-        mapValue.putDouble(valueIndex, Double.NaN);
-        mapValue.putDouble(valueIndex + 1, Double.NaN);
-        mapValue.putLong(valueIndex + 2, 0);
     }
 }
